airsticky/requirements.txt
apache-airflow-providers-snowflake
snowflake-connector-python
snowflake-sqlalchemy
boto3
